refactor(puppeteer-scripts): improve file structure of files automating BambooHR UI interactions
