+++ src/bamboo-hr-interactions/index.ts
@@ -1,0 +1,5 @@
+export { login } from "./login";
+export { addWorkingHoursEntry } from "./working-hours-plugin/add-time-entry";
+export { addWorkingHoursToday } from "./working-hours-plugin/add-working-hours";
+export { openWorkingHoursForm } from "./working-hours-plugin/open-modal";
+export { saveChanges } from "./working-hours-plugin/save-time-entries";
