+++ src/bamboo-hr-interactions/working-hours-plugin/add-time-entry.ts
@@ -1,0 +1,5 @@
+import { Page } from "puppeteer";
+
+export const addWorkingHoursEntry = async (page: Page) => {
+  await page.click(".AddEditEntry__addEntryLink");
+};
