+++ src/bamboo-hr-interactions/working-hours-plugin/open-modal.ts
@@ -1,0 +1,5 @@
+import { Page } from "puppeteer";
+
+export const openWorkingHoursForm = async (page: Page) => {
+  await page.click(".TimeTrackingWidget button");
+};
