+++ src/bamboo-hr-interactions/working-hours-plugin/save-time-entries.ts
@@ -1,0 +1,5 @@
+import { Page } from "puppeteer";
+
+export const saveChanges = async (page: Page) => {
+  await page.click("div[role=contentinfo] button:nth-of-type(1)");
+};
